refactor: extract method from class to new func
